Refactor stat calculation code for clarity
