+++ backend/main/src/test/java/sentiment/stats/StatsTests.java
@@ -52,14 +52,12 @@ public class StatsTests {
   }
 
 	@Test
-	public void testProcessRawReviews() throws Exception {
-    StatsRequest request = new StatsRequest("", "", "2001-01-01T00:00:00.000Z", "2001-01-01T00:00:00.000Z", new IncomingStat[]{});
-
+	public void testRawReviewsCalculation() throws Exception {
     List<Map<String, AttributeValue>> reviews = new ArrayList<Map<String, AttributeValue>>();
     reviews.add(populateReview("test store*test id", "2018-05-21T00:00:00.000Z", "1.0.0", "abcdef"));
     reviews.add(populateReview("test store2*test id2", "2001-04-04T00:00:00.000Z", "0.0.1", "a1b2c3"));
 
-    OutgoingStat<String, String> result = request.processRawReviews(reviews);
+    OutgoingStat<?, ?> result = new RawReviewsCalculation(reviews).calculate();
 
     Map<String, String> reviewsJson = new HashMap<String, String>();
     reviewsJson.put("abcdef", "{\"date\":\"2018-05-21T00:00:00.000Z\",\"reviewHash\":\"abcdef\",\"version\":\"1.0.0\",\"appIdStore\":\"test store*test id\"}");
